Finish Filter screen, pull detail data sucessfully from api

+++ app/src/main/java/com/nguyen/shelter/api/response/FloorPlan.kt
@@ -8,10 +8,10 @@ data class FloorPlan(
     var id: Int?,
 
     @SerializedName("baths")
-    var baths: Int?,
+    var baths: Double?,
 
     @SerializedName("beds")
-    var beds: Int?,
+    var beds: Double?,
 
     @SerializedName("photo_count")
     var photoCount: Int?,
